full dot menu, hooks, refactor code
