+++ frontend/app/components/CardBack.tsx
@@ -13,7 +13,7 @@ import { GestureResponderEvent } from 'react-native';
 import { useTheme } from '../contexts/ThemeContext';
 import SmileyDisplay from './SmileyDisplay';
 import { MotiView } from 'moti';
-import { ContentStep } from './Card';
+import { ContentStep } from '../hooks/useCardFlip';
 import GradientBackground from './GradientBackground';
 
 interface CardBackProps {
@@ -58,86 +58,92 @@ const MeaningContent = ({
   const [showIndicators, setShowIndicators] = useState(false);
 
   return (
-    <View style={styles.contentContainer}>
+    <View style={styles.meaningContentContainer}>
       <View style={styles.titleSection}>
         <Ionicons name="bulb-outline" size={22} color="#FFD700" />
         <Text style={[styles.stepTitle, { color: '#FFD700' }]}>Meaning</Text>
       </View>
 
-      <MotiView
-        from={{ opacity: 0, scale: 0.85, translateY: 24 }}
-        animate={{ opacity: 1, scale: 1, translateY: 0 }}
-        transition={{
-          type: 'timing',
-          duration: 1200,
-          easing: Easing.out(Easing.exp),
-        }}
-        onDidAnimate={(key, finished) => {
-          if (key === 'opacity' && finished) {
-            setTimeout(() => setShowEmojis(true), 100);
-          }
-        }}
-      >
-        <Text style={[styles.cleanText, { color: textColor }]}>{meaning}</Text>
-      </MotiView>
-
-      {showEmojis && (
+      <View style={styles.meaningContent}>
         <MotiView
-          from={{ opacity: 0, scale: 0.85, translateY: 16 }}
+          from={{ opacity: 0, scale: 0.85, translateY: 24 }}
           animate={{ opacity: 1, scale: 1, translateY: 0 }}
           transition={{
-            type: 'spring',
-            damping: 18,
-            stiffness: 120,
-            mass: 0.8,
-            delay: 100,
+            type: 'timing',
+            duration: 1200,
+            easing: Easing.out(Easing.exp),
           }}
           onDidAnimate={(key, finished) => {
-            if (key === 'scale' && finished) {
-              setTimeout(() => setShowIndicators(true), 150);
+            if (key === 'opacity' && finished) {
+              setTimeout(() => setShowEmojis(true), 100);
             }
           }}
         >
-          <>
-            <View style={{ marginTop: 16, marginBottom: 12 }}>
-              <SmileyDisplay smileys={alternativeDepiction} />
-            </View>
-            <View
-              style={{
-                flexDirection: 'row',
-                flexWrap: 'wrap',
-                justifyContent: 'center',
-                marginTop: 8,
-                marginBottom: 16,
+          <Text style={[styles.cleanText, { color: textColor }]}>
+            {meaning}
+          </Text>
+        </MotiView>
+
+        <View style={styles.emojiContainer}>
+          {showEmojis ? (
+            <MotiView
+              from={{ opacity: 0, scale: 0.85, translateY: 16 }}
+              animate={{ opacity: 1, scale: 1, translateY: 0 }}
+              transition={{
+                type: 'spring',
+                damping: 18,
+                stiffness: 120,
+                mass: 0.8,
+                delay: 100,
+              }}
+              onDidAnimate={(key, finished) => {
+                if (key === 'scale' && finished) {
+                  setTimeout(() => setShowIndicators(true), 150);
+                }
               }}
             >
-              {item.category_theme &&
-                item.category_theme.map((category, idx) => (
-                  <View
-                    key={idx}
-                    style={{
-                      backgroundColor: '#FFD70022',
-                      borderRadius: 12,
-                      paddingHorizontal: 10,
-                      paddingVertical: 4,
-                      margin: 2,
-                    }}
-                  >
-                    <Text
-                      style={{
-                        color: '#FFD700',
-                        fontWeight: 'bold',
-                        fontSize: 13,
-                      }}
-                    >
-                      {category}
-                    </Text>
-                  </View>
-                ))}
-            </View>
-          </>
-        </MotiView>
-      )}
+              <>
+                <View style={{ marginTop: 16, marginBottom: 12 }}>
+                  <SmileyDisplay smileys={alternativeDepiction} />
+                </View>
+                <View
+                  style={{
+                    flexDirection: 'row',
+                    flexWrap: 'wrap',
+                    justifyContent: 'center',
+                    marginTop: 8,
+                    marginBottom: 16,
+                  }}
+                >
+                  {item.category_theme &&
+                    item.category_theme.map((category, idx) => (
+                      <View
+                        key={idx}
+                        style={{
+                          backgroundColor: '#FFD70022',
+                          borderRadius: 12,
+                          paddingHorizontal: 10,
+                          paddingVertical: 4,
+                          margin: 2,
+                        }}
+                      >
+                        <Text
+                          style={{
+                            color: '#FFD700',
+                            fontWeight: 'bold',
+                            fontSize: 13,
+                          }}
+                        >
+                          {category}
+                        </Text>
+                      </View>
+                    ))}
+                </View>
+              </>
+            </MotiView>
+          ) : null}
+        </View>
+      </View>
 
       {showIndicators && (
         <MotiView
@@ -201,7 +207,6 @@ const ExamplesContent = ({
 }: ExamplesContentProps) => {
   const [showAllExamples, setShowAllExamples] = useState(false);
   const [showButton, setShowButton] = useState(false);
-  const { colors } = useTheme();
 
   const handleShowMoreExamples = () => {
     setShowAllExamples(true);
@@ -273,7 +278,10 @@ const ExamplesContent = ({
               onPress={handleShowMoreExamples}
               style={[
                 styles.showMoreButton,
-                { backgroundColor: colors.cardBackBackground },
+                {
+                  backgroundColor: 'rgba(255, 255, 255, 0.2)',
+                  borderColor: 'rgba(255, 255, 255, 0.2)',
+                },
               ]}
               activeOpacity={0.8}
             >
@@ -402,7 +410,7 @@ export const CardBack = ({
           style={[
             styles.nextButton,
             {
-              backgroundColor: 'rgba(0, 0, 0, 0.2)',
+              backgroundColor: 'rgba(255, 255, 255, 0.1)',
               borderColor: 'rgba(255, 255, 255, 0.3)',
             },
           ]}
@@ -418,7 +426,7 @@ export const CardBack = ({
           style={[
             styles.backButton,
             {
-              backgroundColor: 'rgba(0, 0, 0, 0.2)',
+              backgroundColor: 'rgba(255, 255, 255, 0.1)',
               borderColor: 'rgba(255, 255, 255, 0.3)',
             },
           ]}
@@ -451,6 +459,15 @@ const styles = StyleSheet.create({
   contentContainer: {
     flex: 1,
     justifyContent: 'flex-start',
+    alignItems: 'center',
+    paddingHorizontal: 12,
+    paddingTop: 40,
+    paddingBottom: 60,
+    maxHeight: '100%',
+  },
+  meaningContentContainer: {
+    flex: 1,
+    justifyContent: 'center',
     alignItems: 'center',
     paddingHorizontal: 12,
     paddingTop: 40,
@@ -565,15 +582,13 @@ const styles = StyleSheet.create({
     marginTop: 16,
     paddingVertical: 12,
     paddingHorizontal: 16,
-    borderRadius: 14,
-    borderWidth: 1.5,
-    borderColor: '#FFD70040',
-    backgroundColor: 'rgba(255, 215, 0, 0.08)',
-    shadowColor: '#FFD700',
-    shadowOffset: { width: 0, height: 3 },
-    shadowOpacity: 0.15,
-    shadowRadius: 6,
-    elevation: 4,
+    borderRadius: 25,
+    borderWidth: 1,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.2,
+    shadowRadius: 4,
+    elevation: 3,
   },
   showMoreContent: {
     flexDirection: 'row',
@@ -586,5 +601,15 @@ const styles = StyleSheet.create({
     marginLeft: 8,
     marginRight: 6,
     letterSpacing: 0.3,
+  },
+  meaningContent: {
+    width: '100%',
+    alignItems: 'center',
+  },
+  emojiContainer: {
+    width: '100%',
+    alignItems: 'center',
+    justifyContent: 'center',
+    minHeight: 120,
   },
 });
